+++ libs/components/manifest/src/generator/testing/fixtures/example-packages/foo/src/lib/foo.class.ts
@@ -111,11 +111,5 @@ export abstract class FooWithStaticPropertiesClass {
 
 /**
  * @docsId docs-id-override
- * @docsIncludeIds FooWithStaticPropertiesClass,     FooClass,FooClass,FooClass
  */
 export class FooWithDocsIdOverrideClass {}
-
-/**
- * @docsIncludeIds
- */
-export class FooWithEmptyDocsIncludeIdsClass {}
