+++ libs/components/manifest/src/generator/testing/fixtures/example-packages/foo/src/lib/foo.module.ts
@@ -1,11 +1,5 @@
 import { ModuleWithProviders, NgModule } from '@angular/core';
 
-import {
-  FooComponent,
-  SkyFooNonStandaloneComponent,
-  SkyFooStandaloneComponent,
-} from './foo.component';
-
 @NgModule({})
 export class FooModule {
   public forRoot(): ModuleWithProviders<FooModule> {
@@ -15,14 +9,3 @@ export class FooModule {
     };
   }
 }
-
-/**
- * This module should automatically generate values for docsIncludeIds based on
- * its exports.
- */
-@NgModule({
-  declarations: [SkyFooNonStandaloneComponent],
-  imports: [FooComponent, SkyFooStandaloneComponent],
-  exports: [SkyFooStandaloneComponent, SkyFooNonStandaloneComponent],
-})
-export class FooWithExportsModule {}
