+++ libs/components/manifest/src/generator/testing/fixtures/example-packages/invalid-docs-id/src/index.ts
@@ -1,7 +1,1 @@
-export {
-  DuplicateDocsId1,
-  DuplicateDocsId2,
-  ReferencingInternalDocsIds,
-  SomeInternalClass,
-  UnrecognizedDocsIncludeIds,
-} from './invalid-docs-id';
+export { DuplicateDocsId1, DuplicateDocsId2 } from './invalid-docs-id';
