+++ libs/components/manifest/src/generator/testing/fixtures/example-packages/invalid-docs-id/src/invalid-docs-id.ts
@@ -7,18 +7,3 @@ export class DuplicateDocsId1 {}
  * @docsId my-duplicate
  */
 export class DuplicateDocsId2 {}
-
-/**
- * @docsIncludeIds foo
- */
-export class UnrecognizedDocsIncludeIds {}
-
-/**
- * @docsIncludeIds SomeInternalClass
- */
-export class ReferencingInternalDocsIds {}
-
-/**
- * @internal
- */
-export class SomeInternalClass {}
