+++ libs/components/manifest/src/generator/utility/get-class.ts
@@ -121,7 +121,6 @@ export function getClass(
     deprecationReason,
     description,
     docsId,
-    docsIncludeIds,
     isDeprecated,
     isInternal,
     isPreview,
@@ -140,7 +139,6 @@ export function getClass(
     deprecationReason,
     description,
     docsId: docsId ?? className,
-    docsIncludeIds,
     filePath,
     isDeprecated,
     isInternal,
