+++ libs/components/manifest/src/generator/utility/get-directive.ts
@@ -117,7 +117,6 @@ export function getDirective(
     deprecationReason,
     description,
     docsId,
-    docsIncludeIds,
     isDeprecated,
     isInternal,
     isPreview,
@@ -136,7 +135,6 @@ export function getDirective(
     deprecationReason,
     description,
     docsId: docsId ?? directiveName,
-    docsIncludeIds,
     filePath,
     isDeprecated,
     isInternal,
