+++ libs/components/manifest/src/generator/utility/get-enum.ts
@@ -52,7 +52,6 @@ export function getEnum(
     deprecationReason,
     description,
     docsId,
-    docsIncludeIds,
     isDeprecated,
     isInternal,
     isPreview,
@@ -65,7 +64,6 @@ export function getEnum(
     deprecationReason,
     description,
     docsId: docsId ?? reflection.name,
-    docsIncludeIds,
     filePath,
     isDeprecated,
     isInternal,
