+++ libs/components/manifest/src/generator/utility/get-function.ts
@@ -18,7 +18,6 @@ export function getFunction(
     deprecationReason,
     description,
     docsId,
-    docsIncludeIds,
     isDeprecated,
     isInternal,
     isPreview,
@@ -31,7 +30,6 @@ export function getFunction(
     deprecationReason,
     description,
     docsId: docsId ?? reflection.name,
-    docsIncludeIds,
     filePath,
     isDeprecated,
     isInternal,
