+++ libs/components/manifest/src/generator/utility/get-interface.ts
@@ -56,7 +56,6 @@ export function getInterface(
     deprecationReason,
     description,
     docsId,
-    docsIncludeIds,
     isDeprecated,
     isInternal,
     isPreview,
@@ -73,7 +72,6 @@ export function getInterface(
     deprecationReason,
     description,
     docsId: docsId ?? decl.name,
-    docsIncludeIds,
     filePath,
     indexSignatures,
     isDeprecated,
