+++ libs/components/manifest/src/generator/utility/get-type-alias.ts
@@ -56,7 +56,6 @@ export function getTypeAlias(
     deprecationReason,
     description,
     docsId,
-    docsIncludeIds,
     isDeprecated,
     isInternal,
     isPreview,
@@ -72,7 +71,6 @@ export function getTypeAlias(
     deprecationReason,
     description,
     docsId: docsId ?? reflection.name,
-    docsIncludeIds,
     filePath,
     isDeprecated,
     isInternal,
