+++ libs/components/manifest/src/generator/utility/get-variable.ts
@@ -16,7 +16,6 @@ export function getVariable(
     deprecationReason,
     description,
     docsId,
-    docsIncludeIds,
     isDeprecated,
     isInternal,
     isPreview,
@@ -29,7 +28,6 @@ export function getVariable(
     deprecationReason,
     description,
     docsId: docsId ?? reflection.name,
-    docsIncludeIds,
     filePath,
     isDeprecated,
     isInternal,
