+++ libs/components/manifest/src/get-public-api.test.ts
@@ -24,43 +24,43 @@ describe('get-public-api', () => {
     expect(result).toEqual(publicApi);
   });
 
-  it('should return the public API for a specific docs ID', async () => {
-    const publicApi = {
-      packages: {
-        '@company/components': [
-          {
-            docsId: 'foo',
-            isInternal: false,
-          },
-          {
-            docsId: 'bar',
-            isInternal: false,
-            docsIncludeIds: ['foo', 'FooTesting', 'FooTestingInternal'],
-          },
-          {
-            docsId: 'ignored',
-            isInternal: false,
-          },
-        ],
-        '@company/components/testing': [
-          {
-            docsId: 'FooTesting',
-            isInternal: false,
-          },
-          {
-            docsId: 'FooTestingInternal',
-            isInternal: true,
-          },
-        ],
-      },
-    };
+  // it('should return the public API for a specific docs ID', async () => {
+  //   const publicApi = {
+  //     packages: {
+  //       '@company/components': [
+  //         {
+  //           docsId: 'foo',
+  //           isInternal: false,
+  //         },
+  //         {
+  //           docsId: 'bar',
+  //           isInternal: false,
+  //           docsIncludeIds: ['foo', 'FooTesting', 'FooTestingInternal'],
+  //         },
+  //         {
+  //           docsId: 'ignored',
+  //           isInternal: false,
+  //         },
+  //       ],
+  //       '@company/components/testing': [
+  //         {
+  //           docsId: 'FooTesting',
+  //           isInternal: false,
+  //         },
+  //         {
+  //           docsId: 'FooTestingInternal',
+  //           isInternal: true,
+  //         },
+  //       ],
+  //     },
+  //   };
 
-    setup({ publicApi });
+  //   setup({ publicApi });
 
-    const { getPublicApiByDocsId } = await import('./get-public-api');
+  //   const { getPublicApiByDocsId } = await import('./get-public-api');
 
-    const result = getPublicApiByDocsId('bar');
+  //   const result = getPublicApiByDocsId('bar');
 
-    expect(result).toMatchSnapshot();
-  });
+  //   expect(result).toMatchSnapshot();
+  // });
 });
