+++ libs/components/manifest/src/utils.test.ts
@@ -7,7 +7,6 @@ describe('utils', () => {
       anchorId: '',
       deprecationReason: '',
       docsId: 'SkyManifestParentDefinition',
-      docsIncludeIds: undefined,
       filePath: '',
       kind: 'directive',
       isDeprecated: false,
@@ -22,7 +21,6 @@ describe('utils', () => {
       anchorId: '',
       deprecationReason: '',
       docsId: 'SkyManifestParentDefinition',
-      docsIncludeIds: ['Foo'],
       filePath: '',
       kind: 'service',
       name: '',
